Set name after parent construct
In php8.1 version of Silverstripe Framework 4.12.7 'name' falls back to autogenerated name from 'title'. This results in name 'Focuspoint' instead of 'FocusPoint'. Result is that the db field is not found, causing an error in asset browser. If name is set after the construct call the issue this is solved.